+++ CHANGELOG.md
@@ -15,6 +15,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 - `transition_duration` property to stack widget.
 - `widget_control` utility function for dynamic widget handling.
 - `text` and `show_text` property to progressbar widget.
+- `graph` widget back.
+
+### Changed
+
+- `v4_18` feature flag for to the `gtk4` crate to enable newer APIs.
 - `content_fit` property to image widget.
 - `can_shrink` property to image widget.
 - `mutations` property to localsignal.